+++ src/app/dataset-view/additional-components/dataset-settings-component/tabs/ingest-configuration/dataset-settings-ingest-configuration-tab.component.ts
@@ -18,7 +18,6 @@ import { takeUntilDestroyed } from "@angular/core/rxjs-interop";
 @Component({
     selector: "app-dataset-settings-ingest-configuration-tab",
     templateUrl: "./dataset-settings-ingest-configuration-tab.component.html",
-    styleUrls: ["./dataset-settings-ingest-configuration-tab.component.scss"],
     changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class DatasetSettingsIngestConfigurationTabComponent extends BaseComponent {
