+++ src/app/dataset-view/additional-components/dataset-settings-component/tabs/ingest-configuration/resolver/dataset-settings-ingest-configuration.resolver.spec.ts
@@ -1,17 +1,24 @@
-import { TestBed } from '@angular/core/testing';
-import { ResolveFn } from '@angular/router';
+/**
+ * Copyright Kamu Data, Inc. and contributors. All rights reserved.
+ *
+ * Use of this software is governed by the Business Source License
+ * included in the LICENSE file.
+ */
 
-import { datasetSettingsIngestConfigurationResolver } from './dataset-settings-ingest-configuration.resolver';
+import { TestBed } from "@angular/core/testing";
+import { ResolveFn } from "@angular/router";
+import { datasetSettingsIngestConfigurationResolverFn } from "./dataset-settings-ingest-configuration.resolver";
+import { DatasetViewData } from "src/app/dataset-view/dataset-view.interface";
 
-describe('datasetSettingsIngestConfigurationResolver', () => {
-  const executeResolver: ResolveFn<boolean> = (...resolverParameters) => 
-      TestBed.runInInjectionContext(() => datasetSettingsIngestConfigurationResolver(...resolverParameters));
+describe("datasetSettingsIngestConfigurationResolver", () => {
+    const executeResolver: ResolveFn<DatasetViewData | null> = (...resolverParameters) =>
+        TestBed.runInInjectionContext(() => datasetSettingsIngestConfigurationResolverFn(...resolverParameters));
 
-  beforeEach(() => {
-    TestBed.configureTestingModule({});
-  });
+    beforeEach(() => {
+        TestBed.configureTestingModule({});
+    });
 
-  it('should be created', () => {
-    expect(executeResolver).toBeTruthy();
-  });
+    it("should be created", () => {
+        expect(executeResolver).toBeTruthy();
+    });
 });
