+++ src/app/dataset-view/additional-components/dataset-settings-component/tabs/scheduling/dataset-settings-scheduling-tab.component.ts
@@ -66,7 +66,6 @@ export class DatasetSettingsSchedulingTabComponent extends BaseComponent {
                 },
             })
             .pipe(takeUntilDestroyed(this.destroyRef))
-
             .subscribe();
     }
