+++ src/app/dataset-view/dataset-view.module.ts
@@ -82,7 +82,7 @@ import { MatTabsModule } from "@angular/material/tabs";
 import { MatChipsModule } from "@angular/material/chips";
 import { DatasetVisibilityModule } from "../common/components/dataset-visibility/dataset-visibility.module";
 import { QuerySharedModule } from "../query/shared/query-shared.module";
-import { IngestConfigurationModule } from "./additional-components/dataset-settings-component/tabs/scheduling/ingest-configuration-form/ingest-configuration.module";
+import { IngestConfigurationModule } from "./additional-components/dataset-settings-component/tabs/ingest-configuration/ingest-configuration-form/ingest-configuration.module";
 import { IngestTriggerModule } from "./additional-components/dataset-settings-component/tabs/scheduling/ingest-trigger-form/ingest-trigger.module";
 import { FlowsViewModule } from "../dataset-flow/flows-view.module";
 import { DatasetViewMenuModule } from "./dataset-view-menu/dataset-view-menu.module";
